pyproject.toml
[tool.poetry]
name = "cloudcart"
version = "0.1.0"
description = ""
authors = ["Shahsad KP <shahsadkpklr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
django-environ = "^0.11.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-tenants = "^3.7.0"
psycopg2 = "^2.9.10"
celery = {extras = ["redis"], version = "^5.4.0"}
gevent = "^24.10.3"
tenant-schemas-celery = "^3.0.0"
pyotp = "^2.9.0"
django-cors-headers = "^4.6.0"
djangorestframework-camel-case = "^1.4.2"
py-webauthn = "^0.0.4"
webauthn = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
